+++ plugin/storage/grpc/shared/plugin.go
@@ -20,7 +20,7 @@ import (
 	"github.com/hashicorp/go-plugin"
 	"google.golang.org/grpc"
 
-	_ "github.com/jaegertracing/jaeger/proto-gen" // force gogo codec registration
+	_ "github.com/jaegertracing/jaeger/pkg/gogocodec" // force gogo codec registration
 	"github.com/jaegertracing/jaeger/proto-gen/storage_v1"
 )
